moved visualization() to a separate module to avoid circular import at module loadtime
